requirements.txt
pyrogram==2.0.106
tgcrypto
motor
pymongo
python-dotenv
aiofiles
humanize
pytz
